se agrego elemento suspendido en registro de usuarios
