+++ src/App.jsx
@@ -4,7 +4,6 @@ import "bootstrap-icons/font/bootstrap-icons.min.css";
 import Inicio from "./components/pages/Inicio";
 import Menu from "./components/commons/Menu";
 import Footer from "./components/commons/Footer";
-import LoginModal from "./components/commons/Login";
 import DetalleProducto from "./components/pages/producto/DetalleProducto";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import RutasProtegidas from "./components/routes/RutasProtegidas";
@@ -12,30 +11,22 @@ import RutasAdmin from "./components/routes/RutasAdmin";
 import { useState } from "react";
 import Registro from "./components/pages/Registro";
 import Pedidos from "./components/pages/pedidos/Pedidos";
+import Login from "./components/pages/Login";
 
 function App() {
   const usuario = JSON.parse(sessionStorage.getItem("usuarioRollingBistro")) || "";
   const [usuarioLogueado, setUsuarioLogueado] = useState(usuario);
-  const [modalIsOpen, setModalIsOpen] = useState(false);
-
-  const openModal = () => {
-    setModalIsOpen(true);
-  };
-
-  const closeModal = () => {
-    setModalIsOpen(false);
-  };
 
   return (
     <>
       <BrowserRouter>
-        <Menu openLoginModal={openModal} usuarioLogueado={usuarioLogueado} setUsuarioLogueado={setUsuarioLogueado}/>
+        <Menu usuarioLogueado={usuarioLogueado} setUsuarioLogueado={setUsuarioLogueado}/>
         <Routes>
-          <Route exact path="/" element={<Inicio usuarioLogueado={usuarioLogueado} openLoginModal={openModal}></Inicio>}></Route>
+          <Route exact path="/" element={<Inicio usuarioLogueado={usuarioLogueado}></Inicio>}></Route>
           <Route
             exact
             path="/detalle/:id"
-            element={<DetalleProducto usuarioLogueado={usuarioLogueado} openLoginModal={openModal}></DetalleProducto>}
+            element={<DetalleProducto usuarioLogueado={usuarioLogueado}></DetalleProducto>}
           ></Route>
            <Route
             exact
@@ -54,13 +45,13 @@ function App() {
           <Route
             exact
             path="/pedidos"
-            element={<Pedidos usuarioLogueado={usuarioLogueado} openLoginModal={openModal}></Pedidos>}
+            element={<Pedidos usuarioLogueado={usuarioLogueado}></Pedidos>}
           ></Route>
           <Route
             exact
             path="/login"
             element={
-              <LoginModal modalIsOpen={modalIsOpen} closeModal={closeModal} setUsuarioLogueado={setUsuarioLogueado}/>
+              <Login setUsuarioLogueado={setUsuarioLogueado}/>
             }
           ></Route>
           {/*<Route path="*" element={<Error404></Error404>}></Route>*/}
