+++ src/helpers/queries.js
@@ -102,20 +102,7 @@ export const borrarUsuarioAPI = async (id) => {
   }
 };
 
-export const inhabilitarUsuarioAPI = async (id) => {
-  try {
-    const resp = await fetch(`${URL_Usuarios}/${id}/disable`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-    return resp;
-  } catch (error) {
-    console.log(error);
-    throw new Error('Error al inhabilitar el usuario desde la API: ' + error.message);
-  }
-};
+
 
 export const crearPedidoAPI = async (pedidoNuevo) => {
   try {
